+++ src/me/martinez/pe/io/BigEndianReader.java
@@ -1,4 +1,4 @@
-package com.company.io;
+package me.martinez.pe.io;
 
 import java.io.IOException;
