+++ src/me/martinez/pe/io/EndianReader.java
@@ -1,4 +1,4 @@
-package com.company.io;
+package me.martinez.pe.io;
 
 import java.io.IOException;
 import java.util.ArrayList;
