+++ src/me/martinez/pe/io/LittleEndianReader.java
@@ -1,4 +1,4 @@
-package com.company.io;
+package me.martinez.pe.io;
 
 import java.io.IOException;
